requirements.txt
streamlit>=1.29.0
pymupdf>=1.23.0
pandas>=2.1.0
openpyxl>=3.1.0
httpx>=0.25.0
python-dotenv>=1.0.0
numpy>=1.24.0
rich>=13.0.0
pydantic>=2.5.0
pytest>=7.4.0

